Migrate over existing tests
