+++ contracts/ado_base/src/modules/hooks.rs
@@ -125,3 +125,30 @@ where
     }));
     process_module_response(mod_resp)
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use andromeda_protocol::error::ContractError;
+    use cosmwasm_std::Response;
+
+    #[test]
+    fn test_process_module_response() {
+        let res: Option<Response> = process_module_response(Ok(Response::new())).unwrap();
+        assert_eq!(Some(Response::new()), res);
+
+        let res: Option<Response> = process_module_response(Err(StdError::generic_err(
+            "XXXXXXX UnsupportedOperation XXXXXXX",
+        )))
+        .unwrap();
+        assert_eq!(None, res);
+
+        let res: ContractError =
+            process_module_response::<Response>(Err(StdError::generic_err("AnotherError")))
+                .unwrap_err();
+        assert_eq!(
+            ContractError::Std(StdError::generic_err("AnotherError")),
+            res
+        );
+    }
+}
